Add filter and apply to gather
