+++ seismicpro/src/survey/survey.py
@@ -594,144 +594,6 @@ class Survey(GatherContainer, SamplesContainer):  # pylint: disable=too-many-ins
     #                       Survey processing methods                        #
     #------------------------------------------------------------------------#
 
-    @staticmethod
-    def _apply(func, df, axis, unpack_args, **kwargs):
-        """Apply a function to a `pd.DataFrame` along the specified axis.
-
-        Parameters
-        ----------
-        func : callable
-            A function to be applied to `df`.
-        df : pd.DataFrame
-            A `DataFrame` to which the function will be applied.
-        axis : {0 or "index", 1 or "columns", None}
-            An axis along which the function is applied:
-            - 0 or "index": apply a function to each column,
-            - 1 or "columns": apply a function to each row,
-            - `None`: apply a function to the `DataFrame` as a whole.
-        unpack_args : bool
-            If `True`, row or column values are passed to `func` as individual arguments, otherwise the whole array is
-            passed as a single arg. If `axis` is `None` and `unpack_args` is `True`, columns of the `df` are passed to
-            the `func` as individual arguments.
-        kwargs : misc, optional
-            Additional keyword arguments to pass to `func` or `pd.DataFrame.apply`.
-
-        Returns
-        -------
-        result : np.ndarray
-            The result of applying `func` to `df`.
-        """
-        if axis is None:
-            args = (col_val for _, col_val in df.iteritems()) if unpack_args else (df,)
-            res = func(*args, **kwargs)
-        else:
-            # FIXME: Workaround for a pandas bug https://github.com/pandas-dev/pandas/issues/34822
-            # raw=True causes incorrect apply behavior when axis=1 and several values are returned from `func`
-            raw = (axis != 1)
-
-            apply_func = (lambda args, **kwargs: func(*args, **kwargs)) if unpack_args else func
-            res = df.apply(apply_func, axis=axis, raw=raw, result_type="expand", **kwargs)
-
-        if isinstance(res, pd.Series):
-            res = res.to_frame()
-        return res.values
-
-    def filter(self, cond, cols, axis=None, unpack_args=False, inplace=False, **kwargs):
-        """Keep only those rows of `headers` where `cond` is `True`.
-
-        Examples
-        --------
-        Keep only traces whose offset is less than 1500 meters:
-        >>> survey = Survey(path, header_index="FieldRecord", header_cols=["TraceNumber", "offset"], name="survey")
-        >>> survey.filter(lambda offset: offset < 1500, cols="offset", inplace=True)
-
-        Parameters
-        ----------
-        cond : callable
-            A function to be applied to `self.headers` to get a filtering mask. Must return a boolean array whose
-            length equals to the length of `headers` and `True` values correspond to traces to keep.
-        cols : str or list of str
-            `self.headers` columns for which condition is checked.
-        axis : {0 or "index", 1 or "columns", None}, optional, defaults to None
-            An axis along which `cond` is applied:
-            - 0 or "index": apply `cond` to each column,
-            - 1 or "columns": apply `cond` to each row,
-            - `None`: apply `cond` to the `DataFrame` as a whole.
-        unpack_args : bool, optional, defaults to False
-            If `True`, row or column values are passed to `cond` as individual arguments, otherwise the whole array is
-            passed as a single arg. If `axis` is `None` and `unpack_args` is `True`, each column from `cols` is passed
-            to the `cond` as an individual argument.
-        inplace : bool, optional, defaults to False
-            Whether to perform filtering inplace or process a survey copy.
-        kwargs : misc, optional
-            Additional keyword arguments to pass to `cond` or `pd.DataFrame.apply`.
-
-        Returns
-        -------
-        self : Survey
-            Filtered survey.
-
-        Raises
-        ------
-        ValueError
-            If `cond` returns more than one bool value for each row of `headers`.
-        """
-        self = maybe_copy(self, inplace, ignore="headers")
-        cols = to_list(cols)
-        headers = pd.DataFrame(self[cols], columns=cols)
-        mask = self._apply(cond, headers, axis=axis, unpack_args=unpack_args, **kwargs)
-        if (mask.ndim != 2) or (mask.shape[1] != 1):
-            raise ValueError("cond must return a single value for each header row")
-        if mask.dtype != np.bool_:
-            raise ValueError("cond must return a bool value for each header row")
-        self.headers = self.headers.loc[mask[:, 0]].copy()  # Guarantee that a copy is set
-        return self
-
-    def apply(self, func, cols, res_cols=None, axis=None, unpack_args=False, inplace=False, **kwargs):
-        """Apply a function to `self.headers` along the specified axis.
-
-        Examples
-        --------
-        Convert signed offsets to their absolute values:
-        >>> survey = Survey(path, header_index="FieldRecord", header_cols=["TraceNumber", "offset"], name="survey")
-        >>> survey.apply(lambda offset: np.abs(offset), cols="offset", inplace=True)
-
-        Parameters
-        ----------
-        func : callable
-            A function to be applied to `self.headers`. Must return a 2d object with shape (`len(self.headers)`,
-            `len(res_cols)`).
-        cols : str or list of str
-            `self.headers` columns for which the function is applied.
-        res_cols : str or list of str, optional, defaults to None
-            `self.headers` columns in which the result is saved. If not given, equals to `cols`.
-        axis : {0 or "index", 1 or "columns", None}, optional, defaults to None
-            An axis along which the function is applied:
-            - 0 or "index": apply a function to each column,
-            - 1 or "columns": apply a function to each row,
-            - `None`: apply a function to the `DataFrame` as a whole.
-        unpack_args : bool, optional, defaults to False
-            If `True`, row or column values are passed to `func` as individual arguments, otherwise the whole array is
-            passed as a single arg. If `axis` is `None` and `unpack_args` is `True`, each column from `cols` is passed
-            to the `func` as an individual argument.
-        inplace : bool, optional, defaults to False
-            Whether to apply the function inplace or to a survey copy.
-        kwargs : misc, optional
-            Additional keyword arguments to pass to `func` or `pd.DataFrame.apply`.
-
-        Returns
-        -------
-        self : Survey
-            A survey with the function applied.
-        """
-        self = maybe_copy(self, inplace)
-        cols = to_list(cols)
-        headers = pd.DataFrame(self[cols], columns=cols)
-        res_cols = cols if res_cols is None else to_list(res_cols)
-        res = self._apply(func, headers, axis=axis, unpack_args=unpack_args, **kwargs)
-        self.headers[res_cols] = res
-        return self
-
     def set_limits(self, limits):
         """Update default survey time limits that are used during trace loading and statistics calculation.
